+++ lambda/custom/index.js
@@ -259,36 +259,6 @@ const NoIntentHandler = {
 	},
 };
 
-const GetUnreadMessagesIntentHandler = {
-	canHandle(handlerInput) {
-		return handlerInput.requestEnvelope.request.type === 'IntentRequest' &&
-			handlerInput.requestEnvelope.request.intent.name === 'ReadUnreadsIntent';
-	},
-	async handle(handlerInput) {
-		try {
-			const {
-				accessToken
-			} = handlerInput.requestEnvelope.context.System.user;
-			const channelNameData = handlerInput.requestEnvelope.request.intent.slots.readunreadschannel.value;
-			const channelName = helperFunctions.replaceWhitespacesFunc(channelNameData);
-
-			const headers = await helperFunctions.login(accessToken);
-			const unreadCount = await helperFunctions.getUnreadCounter(channelName, headers);
-			const speechText = await helperFunctions.channelUnreadMessages(channelName, unreadCount, headers);
-			let repromptText = ri('GENERIC_REPROMPT');
-
-			return handlerInput.jrb
-				.speak(speechText)
-				.speak(repromptText)
-				.reprompt(repromptText)
-				.withSimpleCard(ri('GET_UNREAD_MESSAGES_FROM_CHANNEL.CARD_TITLE'), speechText)
-				.getResponse();
-		} catch (error) {
-			console.error(error);
-		}
-	},
-};
-
 const AddAllToChannelIntentHandler = {
 	canHandle(handlerInput) {
 		return handlerInput.requestEnvelope.request.type === 'IntentRequest' &&
@@ -859,6 +829,8 @@ const {
     AudioControlPlaybackHandler,
     PausePlaybackHandler
 } = require('./handlers/playback')
+
+const { GetUnreadMessagesIntentHandler } = require('./handlers/getUnreadMessages')
 
 const skillBuilder = new Jargon.JargonSkillBuilder({ mergeSpeakAndReprompt: true }).installOnto(Alexa.SkillBuilders.standard());
